// src/insights/components/dashboard-visualization/components/TimeSeries2/types.ts
export type LineChartData = {
  data: LineChartDataSeries;
  label: string;
  min: number;
  max: number;
  dataSetId: string;
};

export type RequestAndPrepareColumnData = (
  params: CommonParams<"TimeSeries"> & {
    dataRange: DataRange | null;
    pointCount: number;
  }
) => Promise<LineChartData[] | null>;

export type LineChartDataWithColor = LineChartData & { color: string };
// ...
export type ChartData = {
  resizeObserver: ResizeObserver | null;
  numtraces: number;
  xAxisLabel: string;
  yAxisLabel: string;
  range: Range | null;
  fullRange: Range;
  dataSetId: string | null;
  datasetLabels: string[];
  chartColumns: LineChartDataWithColor[];
  minimapChartColumns: LineChartDataWithColor[];
  chartSurface: SciChartSurface | null;
  wasmContext: TSciChart | null;
  overview: SciChartOverview | null;
  xySeriesArray: XyDataSeries[];
  minimapXySeriesArray: XyDataSeries[];
};